add some echos, remove file/dir exists errors

+++ init.sh
@@ -4,6 +4,9 @@ source ./Scripts/isWsl.sh
 source ./Scripts/getDistro.sh
 machine=$(./Scripts/machine.sh)
 
+echo "Initializing dotfiles..."
+echo "Environment: $OS"
+
 # apply config
 rm -f \
   ~/.xinitrc \
@@ -20,7 +23,7 @@ else
 fi
 
 
-git clone https://github.com/zbaylin/rofi-wifi-menu.git ~/.config/scripts/rofi-wifi-menu
+[ ! -d "$HOME/.config/scripts/rofi-wifi-menu" ] && git clone https://github.com/zbaylin/rofi-wifi-menu.git ~/.config/scripts/rofi-wifi-menu
 $copy -rf --symbolic-link $(pwd)/.config/* ~/.config/
 $copy -rf --symbolic-link $(pwd)/.ssh/* ~/.ssh/
 
@@ -40,7 +43,7 @@ ln -sf $(pwd)/.ignore ~/.ignore
 ln -sf $(pwd)/.yarnrc ~/.yarnrc
 
 # tmux config
-[ -d "$HOME/.tmux/plugins/tpm" ] && git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
+[ ! -d "$HOME/.tmux/plugins/tpm" ] && git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
 ln -sf $(pwd)/.tmux.conf ~/.tmux.conf
 
 # set up dropbox-stored note taking
@@ -55,12 +58,12 @@ fi
 #rm -rf ~/.oh-my-zsh
 #sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
 # restore zshrc
-mv ~/.zshrc.pre-oh-my-zsh ~/.zshrc
+[ -f ~/.zshrc.pre-oh-my-zsh ] && mv ~/.zshrc.pre-oh-my-zsh ~/.zshrc
 touch ~/.zs
 touch ~/.zsh-update
 touch ~/.zsh-history
 # copy custom omzsh scripts if omzsh exists
-[ -d "$HOME/.oh-my-zsh" ] && mkdir $HOME/.oh-my-zsh
+[ ! -d "$HOME/.oh-my-zsh" ] && mkdir $HOME/.oh-my-zsh
 $copy -rf --symbolic-link $(pwd)/.oh-my-zsh/* ~/.oh-my-zsh/
 ln -sf $(pwd)/.zshrc ~/.zshrc
 
@@ -68,9 +71,11 @@ ln -sf $(pwd)/.zshrc ~/.zshrc
 mkdir -p ~/.config/nvim
 mkdir -p ~/.config/vim/backup_files
 mkdir -p ~/.config/vim/undo_files
-git clone https://github.com/vivaldi-va/nvim-config.git ~/.config/nvim/
+[ ! -d "$HOME/.config/nvim" ] && git clone https://github.com/vivaldi-va/nvim-config.git ~/.config/nvim/
 
 # install node js and utils
-if [ -d "$HOME/.nvm" ]; then
+if [ ! -d "$HOME/.nvm" ]; then
   curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.34.0/install.sh | bash
 fi
+
+echo "Init complete"
